Split up filenamify and filenamify-path into different files
