+++ index.js
@@ -1,44 +1,8 @@
 'use strict';
-const path = require('path');
-const trimRepeated = require('trim-repeated');
-const filenameReservedRegex = require('filename-reserved-regex');
-const stripOuter = require('strip-outer');
+const filenamify = require('./filenamify');
+const filenamifyPath = require('./filenamify-path');
 
-// Doesn't make sense to have longer filenames
-const MAX_FILENAME_LENGTH = 100;
-
-const reControlChars = /[\u0000-\u001f\u0080-\u009f]/g; // eslint-disable-line no-control-regex
-const reRelativePath = /^\.+/;
-
-const filenamify = (string, options = {}) => {
-	if (typeof string !== 'string') {
-		throw new TypeError('Expected a string');
-	}
-
-	const replacement = options.replacement === undefined ? '!' : options.replacement;
-
-	if (filenameReservedRegex().test(replacement) && reControlChars.test(replacement)) {
-		throw new Error('Replacement string cannot contain reserved filename characters');
-	}
-
-	string = string.replace(filenameReservedRegex(), replacement);
-	string = string.replace(reControlChars, replacement);
-	string = string.replace(reRelativePath, replacement);
-
-	if (replacement.length > 0) {
-		string = trimRepeated(string, replacement);
-		string = string.length > 1 ? stripOuter(string, replacement) : string;
-	}
-
-	string = filenameReservedRegex.windowsNames().test(string) ? string + replacement : string;
-	string = string.slice(0, typeof options.maxLength === 'number' ? options.maxLength : MAX_FILENAME_LENGTH);
-
-	return string;
-};
-
-filenamify.path = (filePath, options) => {
-	filePath = path.resolve(filePath);
-	return path.join(path.dirname(filePath), filenamify(path.basename(filePath), options));
-};
+const filenamifyCombined = filenamify;
+filenamifyCombined.path = filenamifyPath;
 
 module.exports = filenamify;
